Order history list has changed

+++ app/src/main/java/com/example/water_nn/presentation/main/history/NewOrderViewModel.kt
@@ -217,8 +217,8 @@ class NewOrderViewModel(
         priceEmptyBottle = (qtyFullBottle - qtyEmptyBottle) * emptyBottlePrice
     }
 
-    private fun calculateTotalPrice(): Double {
-        return priceFullBottle + priceEmptyBottle
+    private fun calculateTotalPrice(): Int {
+        return (priceFullBottle + priceEmptyBottle).toInt()
     }
 
     override suspend fun isOrderDataValid(orderData: OrderData): Boolean {
